move gate parquet with its own module
