+++ rust/tests/data/checkpoint_with_partitions/_delta_log/00000000000000000002.checkpoint.parquet
