+++ rust/tests/data/checkpoint_with_partitions/_delta_log/_last_checkpoint
